requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ngrok==1.7.0
python-dotenv==1.2.2
requests==2.32.5
tmdbsimple==2.9.1
urllib3==2.6.3
Werkzeug==3.1.6
